+++ test/srt/test_nightly_vlm_models_perf.py
@@ -7,6 +7,7 @@ from sglang.srt.utils import kill_process_tree
 from sglang.test.test_utils import (
     DEFAULT_TIMEOUT_FOR_SERVER_LAUNCH,
     DEFAULT_URL_FOR_TEST,
+    _parse_int_list_env,
     generate_markdown_report,
     is_in_ci,
     parse_models,
@@ -57,14 +58,9 @@ class TestNightlyVLMModelsPerformance(unittest.TestCase):
         )
         cls.base_url = DEFAULT_URL_FOR_TEST
 
-        # Bench knobs for bench_one_batch_server (override by env)
-        def _parse_int_list_env(name: str, default_val: str):
-            val = os.environ.get(name, default_val)
-            return [int(x) for x in val.split(",") if x]
-
         cls.batch_sizes = _parse_int_list_env("NIGHTLY_VLM_BATCH_SIZES", "1,1,2,8,16")
-        cls.input_lens = _parse_int_list_env("NIGHTLY_VLM_INPUT_LENS", "1024")
-        cls.output_lens = _parse_int_list_env("NIGHTLY_VLM_OUTPUT_LENS", "16")
+        cls.input_lens = tuple(_parse_int_list_env("NIGHTLY_VLM_INPUT_LENS", "1024"))
+        cls.output_lens = tuple(_parse_int_list_env("NIGHTLY_VLM_OUTPUT_LENS", "16"))
 
     def test_vlm_models_mmmu_performance(self):
         full_report = (
@@ -74,7 +70,6 @@ class TestNightlyVLMModelsPerformance(unittest.TestCase):
             with self.subTest(model=model):
                 process = popen_launch_server_wrapper(self.base_url, model)
                 model_results = []
-
                 try:
                     # Run bench_one_batch_server against the launched server
                     os.makedirs(PROFILE_DIR, exist_ok=True)
